Mirror Message Service on Platform
Ideally, the implementation of MQTT Manager and Message Service would
be placed in common, but for now, this'll do.

Any message handler ("dispatch handlers") are written into tpx/message/dispatch/*.clj,
using the multifn defined therein (interface.clj) as a base.

Basic messaging now works, now's the time to figure out the exact message types we want.
In terms of message formats, the current implementation is based on the following:
{:message/type :some/key
 :message/body {...}
 :message/meta {:mqtt/topic "<UUID>"}}