Add handling of serial output
When the SIP CLI program (connect) runs, it prints out a lot of
information. In order to run the program programatically, we redirect
the I/O to serial ports which we read from and write to.

The output from SIP CLI is significant, with lots of debug
printing. We need a way to read this information, and determine what
has happened, in order to respond and give feedback to users of the
sytem. tpx.ipc.output handles this.