Fix rest of lines in serial output
With the resulting output from SIP, we run across sitations, where
lines are intertwined with each other, and only the final output says
anything about where we are. In such a sitaution, we will run into
problems where lines is not reset.

A check has been added, where we check if this is the final line in a
sequences of lines meant to produce an output. If that is the case,
and we have not found a match, we reset the lines so that we can read anew.