Add network component
When a Teleporter is not connected to the internet, it sets the
network to a pre-configured setting and boots up a webserver. From
there, a user can set up their local network to match what is expected
from the Teleporter and afterwards surf in to the webserver run by the
Teleporter and reset the IP configuration to their choosing.

Large portions of the Network component has been written, as well as
the Webserver component. The Webserver component is started from the
Network component when a dropped internet connection is detected and a
timeout has run its course. Once the IP has been reset, the Webserver
component is meant to shut itself down.

The network namespace will also hold functions to be used by the App,
when changing network configuration.