Add IPC handlers
Added handlers for IPC callback functions. Right now they only log
that something happened.

Did a lot of manual testing with the integration for IPC/serial, and
some bugs were found and ironed out. It looks to be working now, but
the serial I/O is a bit dodgy.

The output from SIP is a bit more varied than assumed, so we need a
way to handle output that seems to be out of bound, without spamming everyone.