Update serial, add handler and output parsing
Small changes to serial, making it a bit smoother to work with.

Added output parsing of BP/SIP. It's a bit shaky, as the output
produced by BP/SIP is quite large.

For the output, we need to handle what happens in there. Added handler
for functions dedicated to handling the output from SIP/BP.