Add playout delay to ipc.command
Added playout delay to commands. The output is highly problematic, as
it gives only the current delay before you put in the new value, and
does not give the newly registered value as any kind of output. In
order to get the new value, you have to run the same command again,
but this time not put in any new value. Getting any kind of output
from this setup that makes sense is a nightmare, and it's better to
change the program to be more friendly towards a program.