Prepare network component for handling changing network config
Currently, this has not been added to init, as it needs testing on actual
zedboards, but the logic appears to be in place and working as intended:
If there is not connection, a webserver will be spun up with default
static IP config. It's important to note that it now only writes to a file,
/tmp/net-eth1, to avoid mishaps when testing on zedboards. Also, currently,
there is no attempt to bring eth1 down or up.

NB! The push is primarily to pull on zedboard for testing.