Grab the MAC from config.edn
The way the command was setup, the MAC that is gotten, is from the
first hop in the network, typically the router. This means the MAC
will change depending on location. Quick and dirty fix on this is to
get a 'MAC' from the config, where we have full control over what it
is supposed to be.