Fix reader problems in dispatch teleporter
In the comment for dispatch/teleporter.clj a function call got nested
in a map. Since it was in the comments, no REPL would choke on it,
whereas an initial startup of a REPL session chokes on it, as the
reader sees that it is invalid syntax.

Quick moving of the function to outside the map fixes the problem.