Update configure-teleporter script with better error handling
Asserts added for the required arguments based on which action was
picked.

Made changes to copy options so that the template path is no longer
needed.

Fixed the wireguard.server.template.conf file with regards to where
the delimiter shows up. Much easier with it at the bottom as you can
just add to the already existing config file.