Fix MQTT publish bug inside a handler
With the MQTT client used in TPX, a bug has been unearthed where the
MQTT client locks itself up when trying to publish a lot of small
messages in quick succession inside a handler for an incoming MQTT
message. This is now offloaded to sync-platform! which has a 5 second
cooldown.