Move gain to left/right input ports and skip left/right channels
While the HW supports gain for left/right channels PER input
source (ie, jack), the GUI is not supporting it (at the
moment). Temporarily removed the gain for 0-3 and instead focused on
left-gain and right-gain (gains per jack).

Also fixed so that the write relay would print out a debug message
with the register in hex, so that it is easier to reason about
together with the HW guys.