Refactor GPIO bitbanging
GPIO bitbanging needs to be able handle send in either 8 + 8 bits, or
8 + 8 + 8 bits (16 or 24 in total).

When a total of 16 bits are sent in, the first array of 8 bits are for the
analog FPGA and the last array of 8 bits are the value to be sent.

When a total of 24 bits are sent in, the first array of 8 bits are for
the analog FPGA, the second array of 8 bits are for one of the chips
on the analog card, and the last array of 8 bits are the value to be
sent to the chip on the analog card.