Add scheduler component
The Scheduler component runs on chime. The heartbeat should ideally be
moved in there as well, in order to have one place where everything is
defined.

For now the overload checks goes in here, but it might be better that
it is in its own namespace in the future, so as to not overload the
scheduler namespace with lots of information that it should probably
not be privy to.

The overloading checks works, and sets both the red lamp on/off, but
also publishes on the broadcast channel for the TP. The logic has been
set up such that it should only trigger on changes.