Add locking mechanism to GPIO bitbang read/write
Given that GPIO will be used by different threads, it is possible to
come into a scenario where one thread wants to write, and another
wants to read, both doing things at the same time. The result would be
a jumbled mess.

In order to fix that, only one read or write is allowed at a time, and
as such a locking mechanism has been introduced that blocks other
reads/writes until the current read/write is finished. Attempts are
done to read/write while waiting. Default is 5 times, and the wait
time per attempt is a Thread/sleep of 10ms. Worst case scenario is
that the current thread trying to read/write is blocked for about 50ms.