Update versions reporting from BP/FPGA to TPX
Due to the slowness of the BP/TPX integration, the data to version
data to send out regarding BP and FPGA never happened as it
should. The primary reason is how slow it is to process the output from
BP, and the secondary reason is that when BP starts up, it does so
before TPX, which means that TPX has a lot of data to go through
before it can even get the answer to the version number for BP and
FPGA.

We solve this by saving the versions in a file, after they have been
read. When TPX first boots up it sets up the reading of the versions,
then loads the (hopefully) previously saved versions, broadcasts
everything and that's it.

In the event that the versions have not been previously loaded, or
that the versions have been changed, the function that handles the
output from the ipc/command function for versions broadcasts again,
but with the updated versions.

At most we have a window of a couple of seconds where there can be a
mismatch between the versions on the server. For now, it's a good
enough compromise.