Change sequence of get-versions in IPC and add path-reset to end of stream
Gather versions first, as the output from init-hw-values! is way too
spammy for the integration with the serial ports to keep up.

The PLC breaks when a jam is ended. Add a path reset at the end of a
stream (when the FPGA reports that the stream has ended) to try and
fix that.